Added new [workspace] -> wspr_cdk_server
