+++ Cargo.toml
@@ -1,3 +1,4 @@
+workspace = { members = ["wspr_cdk_server"] }
 [package]
 name = "wspr_cdk"
 version = "0.0.2"
@@ -20,7 +21,8 @@ chrono = { version = "0.4.38", features = ["serde"] }
 clickhouse = "0.11.6"
 dotenv = "0.15.0"
 reqwest = { version = "0.12.4", features = ["json"] }
-rocket = { version = "0.5.1", features = ["json"] }
 serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
 tokio = { version = "1.37.0", features = ["full"] }
+
+
