+++ src/lib.rs
@@ -1,6 +1,3 @@
-pub mod server;
 pub mod services;
 pub mod state;
 
-#[macro_use]
-extern crate rocket;
