--- src/server/prelude.rs
@@ -1,1 +1,0 @@
-pub mod prelude;
