+++ wspr_cdk_server/Rocket.toml
@@ -1,0 +1,32 @@
+[development]
+address = "localhost"
+port = 8000
+workers = 4
+keep_alive = 5
+read_timeout = 5
+write_timeout = 5
+log = "normal"
+secret_key = "4MsJGFsqvWNqn7zUg99CavhM1Ac2gsgR"
+limits = { forms = 32768 }
+
+[staging]
+address = "0.0.0.0"
+port = 8000
+workers = 4
+keep_alive = 5
+read_timeout = 5
+write_timeout = 5
+log = "normal"
+secret_key = "4MsJGFsqvWNqn7zUg99CavhM1Ac2gsgR"
+limits = { forms = 32768 }
+
+[production]
+address = "0.0.0.0"
+port = 8000
+workers = 4
+keep_alive = 5
+read_timeout = 5
+write_timeout = 5
+log = "critical"
+secret_key = "4MsJGFsqvWNqn7zUg99CavhM1Ac2gsgR"
+limits = { forms = 32768 }
